Making replication factor and number of connections a configurable parameter
As of now NDBench has RF (replication factor) hardcoded and the number of connections to host
is not configured at all taking only the defaults from datastax driver.
These changes allow both replication factor and number of connections to be configured in the
application.properties file while keeping the default values if not specified.

Signed-off-by: Moreno Garcia <moreno@scylladb.com>