CassJavaDriverGeneric: add options to disable use of batch writes
The use of BATCH in CassJavaDriverGeneric allows the driver to write
multiple rows in a single call to `writeSingle`. Its sometimes
desirable, however, to have `writeSingle` to truly write a single row
so that BATCH is not used.

This can cause the row count validation to fail since not every row is
written to every partition. Ad additional configuration has been added
to disable this valdiation.